Task: 3 Create and Display Tasks with AJAX
Create a new controller named "TaskController".
Define methods in "TaskController" to handle the following:
Display all tasks in a paginated manner (GET request).
Store the submitted task data in the database (POST request).
Create corresponding views for each of the methods in the "resources/views/tasks" folder.
Use AJAX to submit the task creation form asynchronously without reloading the page.
Use AJAX to display the list of tasks dynamically without refreshing the entire page when a new task is added.